+++ node/service/Cargo.toml
@@ -95,8 +95,3 @@ try-runtime = [
 	"crab-runtime/try-runtime",
 	"darwinia-runtime/try-runtime",
 ]
-
-dev = [
-	"crab-runtime/dev",
-	"darwinia-runtime/dev",
-]
