+++ runtime/darwinia/src/pallets/identity.rs
@@ -5,9 +5,9 @@ use crate::{weights::pallet_identity::WeightInfo, *};
 
 frame_support::parameter_types! {
 	// Minimum 100 bytes/CRAB deposited (1 MILLI/byte)
-	pub const BasicDeposit: Balance = deposit(1, 258);
-	pub const FieldDeposit: Balance = deposit(0, 66);
-	pub const SubAccountDeposit: Balance = deposit(1, 53);
+	pub const BasicDeposit: Balance = darwinia_deposit(1, 258);
+	pub const FieldDeposit: Balance = darwinia_deposit(0, 66);
+	pub const SubAccountDeposit: Balance = darwinia_deposit(1, 53);
 	pub const MaxSubAccounts: u32 = 100;
 	pub const MaxAdditionalFields: u32 = 100;
 	pub const MaxRegistrars: u32 = 20;
