+++ runtime/darwinia/src/pallets/proxy.rs
@@ -105,12 +105,12 @@ impl InstanceFilter<Call> for ProxyType {
 
 frame_support::parameter_types! {
 	// One storage item; key size 32, value size 8; .
-	pub const ProxyDepositBase: Balance = deposit(1, 8);
+	pub const ProxyDepositBase: Balance = darwinia_deposit(1, 8);
 	// Additional storage item size of 33 bytes.
-	pub const ProxyDepositFactor: Balance = deposit(0, 33);
+	pub const ProxyDepositFactor: Balance = darwinia_deposit(0, 33);
 	pub const MaxProxies: u16 = 32;
-	pub const AnnouncementDepositBase: Balance = deposit(1, 8);
-	pub const AnnouncementDepositFactor: Balance = deposit(0, 66);
+	pub const AnnouncementDepositBase: Balance = darwinia_deposit(1, 8);
+	pub const AnnouncementDepositFactor: Balance = darwinia_deposit(0, 66);
 	pub const MaxPending: u16 = 32;
 }
